requirements.txt
Django==1.11.17
django-cors-headers==2.1.0
django-jsonfield==1.0.1
djangorestframework==3.6.3
pytz==2018.9
